refactor: Move livekit service tools up one directory level
Move tools/services/livekit/ to tools/livekit/ and update all imports.
